+++ src/Export/ExportCategories.php
@@ -4,7 +4,9 @@ declare(strict_types=1);
 
 namespace Omikron\FactFinder\Shopware6\Export;
 
+use Omikron\FactFinder\Shopware6\Export\Data\Entity\CategoryEntity as ExportCategoryEntity;
 use Omikron\FactFinder\Shopware6\OmikronFactFinder;
+use Shopware\Core\Content\Category\CategoryEntity;
 use Shopware\Core\Framework\DataAbstractionLayer\Search\Criteria;
 use Shopware\Core\Framework\DataAbstractionLayer\Search\Filter\EqualsFilter;
 use Shopware\Core\System\SalesChannel\Entity\SalesChannelRepositoryInterface;
@@ -48,5 +50,15 @@ class ExportCategories implements ExportInterface
         $criteria->addFilter(new EqualsFilter(sprintf('customFields.%s', OmikronFactFinder::CMS_EXPORT_INCLUDE_CUSTOM_FIELD_NAME), true));
 
         return $criteria;
+    }
+
+    public function getCoveredEntityType(): string
+    {
+        return CategoryEntity::class;
+    }
+
+    public function getProducedExportEntityType(): string
+    {
+        return ExportCategoryEntity::class;
     }
 }