+++ src/Export/Field/FilterAttributes.php
@@ -8,6 +8,7 @@ use Omikron\FactFinder\Shopware6\Config\ExportSettings;
 use Omikron\FactFinder\Shopware6\Export\PropertyFormatter;
 use Shopware\Core\Content\Product\SalesChannel\SalesChannelProductEntity as Product;
 use Shopware\Core\Content\Property\Aggregate\PropertyGroupOption\PropertyGroupOptionEntity;
+use Shopware\Core\Framework\DataAbstractionLayer\Entity;
 
 class FilterAttributes implements FieldInterface
 {
@@ -25,13 +26,23 @@ class FilterAttributes implements FieldInterface
         return 'FilterAttributes';
     }
 
-    public function getValue(Product $product): string
+    /**
+     * @param Product $entity
+     *
+     * @return string
+     */
+    public function getValue(Entity $entity): string
     {
-        $attributes = $product->getChildren()->reduce(
+        $attributes = $entity->getChildren()->reduce(
             fn (array $result, Product $child): array => $result + array_map($this->propertyFormatter, $child->getOptions()->getElements()),
-            array_map($this->propertyFormatter, $this->applyPropertyGroupsFilter($product))
+            array_map($this->propertyFormatter, $this->applyPropertyGroupsFilter($entity))
         );
         return $attributes ? '|' . implode('|', array_values($attributes)) . '|' : '';
+    }
+
+    public function getCompatibleEntityTypes(): array
+    {
+        return [Product::class];
     }
 
     private function applyPropertyGroupsFilter(Product $product): array