+++ src/Export/Field/Price.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace Omikron\FactFinder\Shopware6\Export\Field;
 
 use Omikron\FactFinder\Shopware6\Export\Formatter\NumberFormatter;
-use Shopware\Core\Content\Product\SalesChannel\SalesChannelProductEntity as Product;
+use Shopware\Core\Framework\DataAbstractionLayer\Entity;
 
 class Price implements FieldInterface
 {
@@ -21,8 +21,13 @@ class Price implements FieldInterface
         return 'Price';
     }
 
-    public function getValue(Product $product): string
+    public function getValue(Entity $entity): string
     {
-        return $this->numberFormatter->format((float) $product->getCalculatedPrice()->getTotalPrice());
+        return $this->numberFormatter->format((float) $entity->getCalculatedPrice()->getTotalPrice());
+    }
+
+    public function getCompatibleEntityTypes(): array
+    {
+        return [Product::class];
     }
 }