+++ src/Export/Field/PriceCurrency.php
@@ -6,6 +6,7 @@ namespace Omikron\FactFinder\Shopware6\Export\Field;
 
 use Omikron\FactFinder\Shopware6\Export\Formatter\NumberFormatter;
 use Shopware\Core\Content\Product\SalesChannel\SalesChannelProductEntity as Product;
+use Shopware\Core\Framework\DataAbstractionLayer\Entity;
 use Shopware\Core\System\Currency\CurrencyEntity;
 
 class PriceCurrency extends Price
@@ -24,8 +25,13 @@ class PriceCurrency extends Price
         return 'Price_' . $this->currency->getIsoCode();
     }
 
-    public function getValue(Product $product): string
+    public function getValue(Entity $entity): string
     {
-        return $this->numberFormatter->format($product->getCalculatedPrice()->getTotalPrice() * $this->currency->getFactor());
+        return $this->numberFormatter->format($entity->getCalculatedPrice()->getTotalPrice() * $this->currency->getFactor());
+    }
+
+    public function getCompatibleEntityTypes(): array
+    {
+        return [Product::class];
     }
 }